Enabled passing on PIDs from the command line to the debugger.
Introduced struct DebuggerStartParameters to pass around
start parameters, removing the need for the engines to
access private members of DebuggerManager. Pass it
to DebuggerManager::startNewDebugger() and move
the GUI parts of that function into DebuggerPlugin,
making it possible to create the struct from command line
parameters. Introduce command line parsing with a few more
-enable-xx options. Introduce warning() to debugger panel
in the CDB engine.

On this occasion, clean out a few trailing whitespaces.

+++ src/plugins/debugger/debuggerdialogs.h
@@ -87,7 +87,7 @@ public:
     explicit AttachExternalDialog(QWidget *parent);
     ~AttachExternalDialog();
 
-    int attachPID() const;
+    qint64 attachPID() const;
 
 private slots:
     void rebuildProcessList();
